Change the no_ack param in SendCodecCommand to an enum.
Instead of passing a bool to IntelHDAStreamBase::SendCodecCommand,
pass an enum class instead.  The strong type makes it much more clear
at the call site whether or not you are requesting an Ack.

Also...  Change the stream setup/disable code to *not* request acks
for the commands sent.  They were supposed to be fire-and-forget in
the first place, but somehow this got flubbed...  I wonder if I would
have been less likely to screw this up if the parameter had been a
self documenting enum instead of a simple bool :P

Change-Id: Iff4135bf69cc762c6157d26a2938bef5f28bfc0e