Make realtek codecs pay attention to jack detect override
Pay attention to the jack detect override bit set by the BIOS in the
realtek codec driver.  Just because a pin complex is capable of
performing presence detection, it does not mean that the device
manufacturer has provided the circuitry to do so.  For example, the
built-in speakers on Acer 12 are connected to pins in the codec which
could (in theory) do presence detection, but since the speakers are
hard-wired at the board level, the BIOS sets the jack detect override
bit in the config defaults for the pin complex.

Change-Id: I72a03bf463ad682042f39beee09b17fa22d7131a