Add FX, a demonstration guitar effects processor.
Add a demo/example app which binds to an audio input driver directly,
and feeds output to the system mixer.  In the process, it applies a
configurable amount of preamp stage gain as well as some configurable
reverb/echo.

Mostly, this was just for fun.  Contributions are more than welcome.
Some nice things to add would be...

1) A nice GUI UI.
2) Nifty new effects.
3) The ability to order effects chains arbitrarily.
4) The ability to persist and restore effect chains.
5) The ability to process multiple instrument inputs simultaneously
   (each with their own effects, of course)

Change-Id: I0c222b84b61f7543977d26fb7a8e5089e7e1951b