Add GET_FORMATS support to codec utils.
Allow codec implementations who use the codec utils library to tell
their base class which audio formats the do and do not support.
Filter SET_FORMAT requests against this list of supported formats and
automatically deny requests which are not in the list of supported
formats.  Also implement a generic handler for GET_FORMATS so that
codec clients do not have to.

Change-Id: I771658177b4f8027954b2164a19bee74b8f46829