[audio] Rework flags in MediaPacket structures
Instead of using a bool per flag in a MediaPacket, use a uint32 along
with some flag constants instead.

Define flags for the existing bools, and add two more (droppable and
discontinuous).  Add some documentation for all of the flags as well
as a short description of what the no-pts constant is all about.

Update existing code to use the new flags encoding.

Note: I spent some time talking to George about the state of FIDL2
work.  While the concept does not currently exist in FIDL/FIDL2, he is
open to adding a formal flags concept to the IDL and generating some
nice-to-use accessors for flags (or perhaps even bitfield) patterns.
If this ends up happening, I will gladly come back here and eliminate
all of the &, | operators and replace them with the generated
accessors.

Change-Id: I0b1334cf395e91896240cc4e9b515db63cf911a9