[media] cache min_delay to optimize audio playback
The GetMinDelay value is static, once the media format
has been set. Because the format is set at stream creation
time, we can retrieve the minimum delay at this time rather
than making an explicit FIDL call every time. This is a
valuable optimization because currently the client calls
GetMinDelay upon every buffer submission.

On debug builds our library DCHECKs if the requested audio
sample rate exceeds the maximum rate supported by the audio
renderer. We previously enforced an incorrect limit; with
this CL we fix this: 48 kHz (not 96 kHz) is the current max
audio output sample rate for the underlying audio renderer.

With this changelist, the media client library sets the
audio renderer's output gain to 0.0 dB when playback
starts (upon the client's first call to stream_write),
and it sets the gain to AudioRenderer::kMutedGain when
playback stops (when the client calls stream_free).

In the absence of device enumeration code, this library
exposes a single audio output -- the default device. This
changelist contains cleanup related to those constants.

Change-Id: I41583842f58c35f8c5c39b89e0d9e06b262dad5e