netconnector(mdns): publish _fuchsia._tcp. instance by unique name
netconnector proper uses the mDNS service to publish the local
Fuchsia device as an instance of the service _fuchsia._tcp. Prior to
this CL, the instance name it used to publish itself was the original
host name obtained via gethostname().

If the host name is found to be in conflict during the address probe,
publishing this instance would probably fail, because another fuchsia
device is already publishing that instance.

This CL causes netconnector to wait until the address probe is
complete and then publishing using the unique host name produced by
the probing sequence.

Now if two fuchsia machines are given the same host name, one will
use a modified name (e.g. "foo 2" instead of "foo"), and the same
device will publish itself as instance "foo 2" of the _fuchsia._tcp.
service. When fuchsia devices search for nearby fuchsia devices,
they'll see foo and foo 2, and both names will resolve to their
respective correct IP addresses.

Change-Id: Ia3d1db3c240a1bbd47c5df851429871bbd67b8a9