[appmgr] Initial work towards supporting introspection.
Appmgr now publishes a /root_info_experimental directory containing some
introspection information about the jobs and application processes which
it knows about.

The directory structure looks like this:
  /root_info
    /<job label>
      /<process label>
        /process - prints the pid of the application
        /url     - prints the url of the application
        /svc     - the exported service directory for the application
          /my::Service - the actual service which is exported

We also publish information for child jobs under /info_experimental in
the namespace of applications started within the containing job.

This implementation is pretty hacky but it's enough for us to start
playing with introspection a little bit while we work towards replacing
appmgr with a more carefully considered component model.

Limitations:

- only jobs and processes managed by appmgr are listed since appmgr
  doesn't retain enough information about apps started through app runners
  to support introspection
- the directory structure is very simplistic, we currently don't try to
  ensure that labels are unique or even valid filenames
- some programs (like /bin/listen) don't actually serve services over
  their directory and don't listen for incoming requests (or close their
  channel), so recursively traversing the info tree may hang when those
  nodes are encountered
- there's no security
- all of this is subject to change...

Change-Id: Ica53faabe3bc536c0c3ade90ce26ad6787b0c934