[rust][zircon] increase call timeout significantly
The bug is tracked in ZX-1324. We're hitting the timeout case far too often
in CQ. As far as we can see there is no bug in the crate code here, and there
are suggestions of switching to INFINITE instead. This is a half-way house
that should be long enough that some kind of host/lower layer stall does not
cause us to fail, and any such stall that is caused by a bug should be
tracked and debugged elsewhere.

Change-Id: I4006071fcab75bda589d0b6240911ecbed7d11d3