[bt][gap] Fix bug in advertising interval computation
Due to an overflow in the timeslice-to-millisecond conversion,
LegacyLowEnergyAdvertiser always sets the advertising interval to
10.24 seconds, which noticeably impairs discovery and connection
establishment time as a LE peripheral.

The cause is missing parentheses in the widening conversion (following
a C-style cast) which happened after an addition overflow. Fixed this by
using a C++ style static_cast around the correct operand.

Also broke down LegacyLowEnergyAdvertiser::StartAdvertising into helper
functions.

NET-270

Change-Id: I13533d65ba111ebf4f70ec38c02cc6bbeff7f664