[wlanstack] Support optional join-by-BSSID
Before:
A typical network manager allows a user to choose (E)SSID, but not BSSID.
Given the ESSID, it is the network manager's role to single out,
supposedly the best, BSSID. This imposes a challenge in development where
the environment has many equally good BSSIDs for the same ESSID, and it
is desired to choose a particular BSSID. This is so especially when
the notion and the implementation of "the best" BSSID is not mature.

After:
default.json employs an optional configuration, "BSSID:". Without this or
with this as an empty string, the behavior is same as before.
When this field is present as a non-empty string, the wlanstack will
attempt to get associated with this particular BSSID only. In this case,
it tends to take longer to get associated.

Change-Id: I45f38120686d7da028e8bc42d44f2c5f2b0cb76a