[netstack] Hotfix for multiple DHCP client instantiations for WLAN NIC
This fix is meant to be a quick work-around for a DHCP regression
problem for WLAN NICs.

This CL distinguishes the feature of the NIC, and apply a different DHCP Client
workflow.

DHCP client should run on the network interface
when (a) the interface goes UP,
     (b) its Link status goes UP,
     (c) the NIC is configured not to use static IP address but to use DHCP

(b) signal is available to WLAN NIC but not to wired Ethernet.

NET-298

Change-Id: I67f0741b4ed589d6401f802f5ca2dacdd39325f2